requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
langchain==0.3.26
langgraph==0.6.8
langchain-openai==0.3.27
openai==1.93.0
httpx==0.26.0
python-dotenv==1.0.1
pydantic==2.7.4

